+++ docs/.vitepress/config.ts
@@ -92,8 +92,6 @@ export default {
             { text: "Overview", link: "/getting-started/overview" },
             { text: "Install", link: "/getting-started/install" },
             { text: "Dependencies", link: "/getting-started/dependencies" },
-            { text: "Install Options", link: "/getting-started/options" },
-            { text: "Install in VM (KVM)", link: "/getting-started/vm-install" },
             { text: "Update", link: "/getting-started/update" },
           ],
         },
@@ -102,8 +100,7 @@ export default {
           collapsed: true,
           items: [
             { text: "Preserve Config & Customize", link: "/configuration/preserve-config" },
-            { text: "Auto Setup & Update", link: "/configuration/auto-setup" },
-            { text: "Use on Other Distros", link: "/configuration/distros" },
+            { text: "Use on other Distros", link: "/configuration/distros" },
             { text: "Monitor Setup", link: "/configuration/monitor-setup" },
             { text: "Hyprland + NVIDIA", link: "/configuration/hypr-nvidia" },
             { text: "Switch SDL (X11/Wayland)", link: "/configuration/xwayland" },
@@ -118,7 +115,6 @@ export default {
             { text: "Screenshots", link: "/usage/screenshots" },
             { text: "Game Mode", link: "/usage/game-mode" },
             { text: "Wallpapers", link: "/usage/wallpapers" },
-            { text: "Power & Screenlock", link: "/usage/power-lock" },
           ],
         },
         {