+++ setup/setup-arch.sh
@@ -28,6 +28,7 @@ packages=(
     "fastfetch"
     "xdg-desktop-portal-gtk"
     "eza"
+    "nautilus"
     "python-pip"
     "python-gobject"
     "python-screeninfo"
@@ -56,7 +57,6 @@ packages=(
     "rofi-wayland"
     "polkit-gnome"
     "zsh"
-    "zsh-completions"
     "fzf"
     "pavucontrol"
     "papirus-icon-theme"
@@ -118,21 +118,13 @@ _installYay() {
 }
 
 _installPackages() {
-    toInstall=()
     for pkg; do
         if [[ $(_isInstalled "${pkg}") == 0 ]]; then
             echo ":: ${pkg} is already installed."
             continue
         fi
-        toInstall+=("${pkg}")
+        yay --noconfirm -S "${pkg}"
     done
-    if [[ "${toInstall[@]}" == "" ]]; then
-        return
-    fi
-    if [[ ! ${toInstall[@]} == "cargo" ]]; then
-        printf "Package not installed:\n%s\n" "${toInstall[@]}"
-    fi
-    yay --noconfirm -S "${toInstall[@]}"
 }
 
 # Header
@@ -177,13 +169,16 @@ fi
 _installPackages "${packages[@]}"
 
 # Oh My Posh
-sudo curl -s https://ohmyposh.dev/install.sh | bash -s
+curl -s https://ohmyposh.dev/install.sh | bash -s
 
 # Cargo
-cargo install -q matugen
-cargo install -q wallust
+echo ":: Installing packages with cargo (this can take a while...)"
+cargo install matugen
+cargo install wallust
 
 # ML4W Apps
+echo ":: Installing the ML4W Apps"
+
 ml4w_app="com.ml4w.welcome"
 ml4w_app_repo="dotfiles-welcome"
 echo ":: Installing $ml4w_app"
@@ -210,7 +205,7 @@ echo ":: Installing $ml4w_app"
 bash -c "$(curl -s https://raw.githubusercontent.com/mylinuxforwork/$ml4w_app_repo/master/setup.sh)"
 
 # Flatpaks
-flatpak install flathub com.github.PintaProject.Pinta
+flatpak install -y flathub com.github.PintaProject.Pinta
 
 # Fonts
 sudo cp -rf $SCRIPT_DIR/fonts/FiraCode /usr/share/fonts