+++ setup/setup-fedora.sh
@@ -35,6 +35,7 @@ packages=(
     "gtk4"
     "libadwaita"
     "fuse"
+    "nautilus"
     "ImageMagick"
     "jq"
     "xclip"
@@ -65,7 +66,6 @@ packages=(
     "bibata-cursor-themes"
     "fontawesome-fonts"
     "dejavu-fonts-all"
-    "flatpak"
     "NetworkManager-tui"
     "nwg-dock-hyprland"
 )
@@ -96,19 +96,13 @@ _isInstalled() {
 }
 
 _installPackages() {
-    toInstall=()
     for pkg; do
         if [[ $(_isInstalled "${pkg}") == 0 ]]; then
             echo "${pkg} is already installed."
             continue
         fi
-        toInstall+=("${pkg}")
+        sudo dnf install --assumeyes --skip-unavailable "${pkg}"
     done
-    if [[ "${toInstall[@]}" == "" ]]; then
-        return
-    fi
-    printf "Package not installed:\n%s\n" "${toInstall[@]}"
-    sudo dnf install --assumeyes --skip-unavailable "${toInstall[@]}"
 }
 
 # Header
@@ -160,14 +154,16 @@ gpgkey=https://repo.charm.sh/yum/gpg.key' | sudo tee /etc/yum.repos.d/charm.repo
 sudo yum install --assumeyes gum
 
 # Oh My Posh
-sudo curl -s https://ohmyposh.dev/install.sh | bash -s
+curl -s https://ohmyposh.dev/install.sh | bash -s
 
 # Cargo
-cargo install -q matugen
-cargo install -q wallust
-cargo install -q eza
+echo ":: Installing packages with cargo (this can take a while...)"
+cargo install matugen
+cargo install wallust
+cargo install eza
 
 # Pip
+echo ":: Installing packages with pip"
 sudo pip install hyprshade
 sudo pip install pywalfox
 sudo pywalfox install
@@ -175,6 +171,8 @@ sudo pip install screeninfo
 sudo pip install waypaper
 
 # ML4W Apps
+echo ":: Installing the ML4W Apps"
+
 ml4w_app="com.ml4w.welcome"
 ml4w_app_repo="dotfiles-welcome"
 echo ":: Installing $ml4w_app"
@@ -201,7 +199,7 @@ echo ":: Installing $ml4w_app"
 bash -c "$(curl -s https://raw.githubusercontent.com/mylinuxforwork/$ml4w_app_repo/master/setup.sh)"
 
 # Flatpaks
-flatpak install flathub com.github.PintaProject.Pinta
+flatpak install -y flathub com.github.PintaProject.Pinta
 
 # Grimblast
 sudo cp $SCRIPT_DIR/scripts/grimblast /usr/bin